dict-req-data, dictionary.asm, macro.asm, and printBox.asm combined w/ bug fixes
Running this file should read a dictionary file, load its words into memory, scramble a 9 letter word, and output it in the box format it. 
You can use this for your own parts of the program that might depend upon these parts.

NOTE: your MIPS .jar needs to be in the same directory as the dictionary files in order for it to be able to access them.